### 2019-10-28 #1838 @whedon
[PRE REVIEW]: splot: visual analytics for spatial statistics
**Submitting author:** @slumnitz (<a href="http://orcid.org/0000-0002-7007-5812">Stefanie Lumnitz</a>)
**Repository:** <a href="https://github.com/pysal/splot" target ="_blank">https://github.com/pysal/splot</a>
**Version:** v1.1.1
