### 2019-10-28 #1838 @slumnitz
Suggestions for potential reviewers:
* [jorisvandenbossche](https://github.com/jorisvandenbossche) - geopandas
* [gboeing](https://github.com/gboeing) - osmnx
* [ResidentMario](https://github.com/ResidentMario) - geoplot
