### 2019-10-30 #1838 @leouieda
Hi @slumnitz thank you for the reviewer suggestions! I just opened a PR to fix a citation formatting issue on the paper. Once that is merged, we can ask @whedon to rebuild the PDF.
