### 2019-10-30 #1838 @leouieda
:wave: Hi @jorisvandenbossche @gboeing would you like to review this submission for JOSS? "splot: visual analytics for spatial statistics" by @slumnitz (Stefanie Lumnitz) *et al*. Here are links to the [submitted paper](https://github.com/openjournals/joss-papers/blob/joss.01838/joss.01838/10.21105.joss.01838.pdf) and [software repository](https://github.com/pysal/splot). Please see [our documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for more information on what a JOSS review is like and feel free to ask any questions you may have (either here or privately by email).
