### 2019-10-30 #1838 @jorisvandenbossche
@leouieda I am happy to review this submission.

To disclose a potential conflict of interest: I acted as co-mentor during Stefanie's GSOC in 2018 while she worked on `splot` (as geopandas maintainer, to give feedback on the usage of geopandas in splot). However, I would describe my role as very limited (I was not involved in the direction or design of the package, but just gave some code feedback, and I also have no commit to the package), and so see myself able to do an impartial review. But let the final decision to you.
