### 2019-11-04 #1838 @leouieda
@jorisvandenbossche thank you for the disclosure, it's much appreciated! I'll follow Dan's suggestion and try to find someone else for the time being so we can avoid the conflict. If I can't find anyone suitable, I'll contact you again to see if you're still available/willing.
