### 2019-11-04 #1838 @leouieda
:wave: Hi @ResidentMario would you like to review this submission for JOSS? "splot: visual analytics for spatial statistics" by @slumnitz (Stefanie Lumnitz) et al. Please see the comments above for more details.
