### 2019-11-04 #1838 @gboeing
I would normally be happy to assist with this, but I don't think I have the capacity to take it on right now. Best of luck -- I saw splot recently and it looks like a cool project!
