### 2019-11-07 #1838 @leouieda
:wave: @kbevers would you like to review this submission for JOSS? "splot: visual analytics for spatial statistics" by @slumnitz (Stefanie Lumnitz) et al. Please see the comments above for more details.
