### 2019-11-10 #1838 @martinfleis
@leouieda If you're still looking for a reviewer, I'd be happy to do it. I've already used PySAL and splot in my own research, so this is very close to my interests. I guess I'll spend time exploring the capabilities of splot anyway.
