### 2019-11-13 pysal/splot#88 @martinfleis
Installation instructions; pip install fails on macOS
Hi,

as part of https://github.com/openjournals/joss-reviews/issues/1882 I am trying to simply follow your installation docs and initially I did not succeed. (under macOS 10.15)

```
conda create -n splot python
conda activate splot
pip install splot
```

```
...
Collecting pyproj
  Downloading https://files.pythonhosted.org/packages/8c/7e/c34acb5abca1916ba0a294df4a90935d6d1cfc446ad1c076986bbf0118d3/pyproj-2.4.1.tar.gz (462kB)
     |████████████████████████████████| 471kB 2.8MB/s
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
   command: /Users/martin/anaconda3/envs/splot/bin/python /Users/martin/anaconda3/envs/splot/lib/python3.8/site-packages/pip/_vendor/pep517/_in_process.py get_requires_for_build_wheel /var/folders/d0/qyf0yvzd0n9ctn2wr3r4l1vw0000gn/T/tmp889iudeu
       cwd: /private/var/folders/d0/qyf0yvzd0n9ctn2wr3r4l1vw0000gn/T/pip-install-4pkixl3a/pyproj
  Complete output (1 lines):
  ERROR: Minimum supported proj version is 6.2.0, installed version is 5.2.0.
ERROR: Command errored out with exit status 1: /Users/martin/anaconda3/envs/splot/bin/python /Users/martin/anaconda3/envs/splot/lib/python3.8/site-packages/pip/_vendor/pep517/_in_process.py get_requires_for_build_wheel /var/folders/d0/qyf0yvzd0n9ctn2wr3r4l1vw0000gn/T/tmp889iudeu Check the logs for full command output.
```

I am aware that you are saying that it is compatible with 3.5 and 3.6 only (why? because of this issue?). Using `conda create -n splot python=3.6` it goes well. But it can be confusing.

I am perfectly aware that this is not `splot` issue, but `pyproj`, being dependency of GeoPandas. As installation of `geopandas` from `pip` is notoriously problematic (http://geopandas.org/install.html), I would welcome a bit more information in your docs what to do. I can install `geopandas` from `conda-forge` and then `splot` from `pip`, which works perfectly, but this information could be in the documentation.

I see that splot is on conda-forge in the same version as on PyPI, why you don't mention this option as well? `conda install -c conda-forge splot` works smoothly.

I would recommend adding `conda-forge` option to documentation and link to` geopandas` docs for `pip` troubleshooting (as the issue is caused by `geopandas`).

Also, I do not understand '_splot supports python 3.5 and 3.6 only_' note as from conda-forge it comes with 3.8.
