### 2019-11-13 pysal/splot#88 @martinfleis
One more thing.

>  Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

It is handled by pip or conda, but it might be nice to list dependencies you require so we know what you are installing.
