### 2019-11-13 5020b59
Fix docstring
