### 2019-11-13 pysal/splot#88 @martinfleis
Now I see that all of it is in Readme. Maybe it is worth to put it to docs as well?
