### 2019-11-13 pysal/splot#90 @martinfleis
Usage in readme is a fragment
Hi,

this is a very minor thing, but Usage section in the readme is a strange fragment only.

> splot supports many different https://github.com/pysal/splot/blob/master/notebooks/mapping_vba.ipynb

I would assume that it points me to https://github.com/pysal/splot/tree/master/notebooks, or that is not the intention with examples?

cc openjournals/joss-reviews#1882
