### 2019-11-17 pysal/splot#91 @martinfleis
DOC: comments to examples 
Hi,
few comments to examples as a part of openjournals/joss-reviews#1882.

All of them are nice and detailed, covering a lot of options `splot` offers, so I have just a couple minor remarks. None of them are critical for the review, but they might help.

1. It is unclear where to find them. There are three examples on the homepage of docs, but 5 in the notebooks folder of the repo. Nowhere in the docs is the link to the rest. Documentation nor readme do not indicate that there are any other resources to check.

2. Space-time visualisations (giddy) is missing a description. There are just a minor comments in the code. It would be good to explain what is going on, why and how to interpret resulting plots.

3. In `mapping_vba`, I assume you wanted to do `from splot.mapping import shift_colormap` instead if `from splot._viz_utils import shift_colormap` as `_viz_utils` is  private.

4. There are some typos here and there across notebooks.
