### 2019-11-17 pysal/splot#92 @martinfleis
JOSS: missing figure captions
Hi,
I believe this is the last issue I am opening as a part of https://github.com/openjournals/joss-reviews/issues/1882.

Figures in you paper are missing captions. I see that they are in the `paper.md`, but they did not show up in pdf. I guess it is an issue with markdown syntax. Can you check that?
