### 2019-11-17 #1882 @martinfleis
Hi,

I would like to close up my review. Majority of `splot`'s feature is currently focused on visualising the data from various PySAL sub-modules, which is extremely helpful in understanding their outputs. However, `splot` goes further than that (and has plans to do even more). A prime example is Value-by-alpha plotting feature. The default visual style of `splot` is very polished, and I would also say that there is no need to touch the defaults and plots can go directly to publications. That is not very common for scientific software.

During my review, I have opened a couple of issues (all linked above). Most of them are just recommendations, rather than critical issues. Before I tick Installation box, I would like to see pysal/splot#88 fixed as there are currently two versions of instructions, with those in the documentation failing. The second issue I would like to see resolved to tick the final box is pysal/splot#92, but that seems to be just a syntax issue.

To sum it up, `splot` is a valuable part of Python geospatial world and I will be watching closely its future development. My last recommendation is to have a detailed discussion with `geoplot` (already started in pysal/splot#21) to agree on the scope of both projects and avoid duplication of effort in the future.
