### 2019-11-19 #1882 @ResidentMario
I'm done with my review. I don't have any hard blocking issues to add to those opened by @martinfleis.

But I will say that I have some qualms with the uneven documentation in `splot`. For example, one of the three tutorial notebooks, titled [Exploratory Analysis of Spatial Data](https://nbviewer.jupyter.org/github/pysal/splot/blob/master/notebooks/esda_morans_viz.ipynb), mentions the following statistical concepts without giving much explanation or linking the user to a reference on what these things are: Moran's I, local Moran, Donatns [sic?] variable, bivariate Moran statistic. Furthermore, [the tutorials](https://splot.readthedocs.io/en/latest/index.html#) only cover a small subset of the library functions: there are many plotters documented at some length in the API Reference which do not appear in the any of the tutorials.

I do not think this is a blocking issue because the documentation is still within the acceptable range for JOSS (all of the relevant methods are covered by the API Reference, which doesn't leave any fields unexplained and is generally very well-packed with example images), but I feel that it's a definite area for improvement for the package as a whole.

I have left more general feedback about documentation with the broader PySAL team in the [PySAL Gitter chat](https://gitter.im/pysal/pysal?at=5dd1dfd7e75b2d5a19efb7cc).
