### 2019-11-25 #1882 @leouieda
Hi @martinfleis @ResidentMario thank you for your reviews! :tada:

@ResidentMario that is a good point about the tutorial coverage and I agree that it's not a requirement for JOSS to have full coverage of tutorials. However, the citation/linking of statistical concepts is highly encouraged as I see scientific software documentation (particularly for JOSS) as a major part of the scholarship. @slumnitz could you please go through the tutorials and add references as you see necessary?

@martinfleis @slumnitz please let me know when the install and figure caption issues are resolved.
