### 2019-11-28 #1882 @slumnitz
@martinfleis @ResidentMario @leouieda thank you for your thorough review, comments and great suggestions! I will implement them in the next couple of weeks or as soon as I can and let you know when everything is done.
