### 2019-12-13 e7ba101
addressing pysal/pysal#1145
