### 2019-12-17 05da21a
ensure datasets are downloaded in .travis.yml
