### 2020-01-05 c64152f
BUG: Fix breakage due to mapclassify deprecation
