### 2020-01-07 #1882 @slumnitz
> @slumnitz - any news on this?

@danielskatz I will start working on it slowly again now. It got a bit delayed due to snowmageddon and a couple of electricity outages here in BC, Canada, which are expected to ease in the next wo weeks.
