### 2020-01-08 4cb01af
[DOC] update installation instruction
* include conda forge in 'install.rst'
* include dependencies
* add troubleshooting paragraph
