+++ doc/installation.rst
@@ -3,23 +3,70 @@
 Installation
 ============
 
-splot supports python `3.5`_ and `3.6`_ only. Please make sure that you are
-operating in a python 3 environment.
+Installing dependencies
+-----------------------
 
-Installing released version
----------------------------
+`splot` is compatible with Python `3.6`_ and `3.7`_ and
+depends on GeoPandas 0.4.0 or later and matplotlib 2.2.2 or later.
+Please make sure that you are operating in a Python 3 environment.
 
-giddy is available on the TBA. Therefore, you can either
-install directly with `pip` from the command line::
+splot also uses
 
-  pip install splot
+* numpy
+* seaborn
+* mapclassify
+* Ipywidgets
+
+Depending on your spatial analysis workflow and the PySAL objects
+you would like to visualize, splot relies on:
+
+PySAL 2.0
+
+or separate packages found in the PySAL stack:
+
+* esda
+* libpysal
+* spreg
+* giddy
 
 
-or download the source distribution (.tar.gz) and decompress it to your selected
-destination. Open a command shell and navigate to the decompressed folder.
-Type::
+Installing the newest release
+-----------------------------
 
-  pip install .
+There are two ways of accessing splot. First, splot is installed with
+the PySAL 2.0 metapackage through:
+
+```$ pip install -U pysal```
+
+or 
+
+```$ conda install -c conda-forge pysal```
+
+Second, splot can be installed as a separate package. If you are
+using Anaconda, install splot via the conda utility:
+
+```$ conda install -c conda-forge splot```
+
+Otherwise you can install splot from PyPI with pip:
+
+```$ pip install splot```
+
+
+Trouble shooting
+----------------
+Most common installation errors are due to splot's dependency on GeoPandas.
+
+It often helps to first install GeoPandas separately from conda-forge with:
+
+```$ conda install geopandas```
+
+before installing splot (preferably also from conda, alternatively from pip).
+
+For more information on troubleshooting the installation of GeoPandas wiht pip, see the `GeoPandas`_ docuemntation.
+
+It is also possible to install splot with a later Python version (>3.7)
+through the separate installation of GeoPandas or through installation wiht conda-forge.
+
 
 Installing development version
 ------------------------------
@@ -36,7 +83,9 @@ your fork. By making changes
 to your local clone and submitting a pull request to `pysal/splot`_, you can
 contribute to the splot development.
 
-.. _3.5: https://docs.python.org/3.5/
 .. _3.6: https://docs.python.org/3.6/
+.. _3.7: https://docs.python.org/3.7/
+.. _GeoPandas: http://geopandas.org/install.html
 .. _pysal/splot: https://github.com/pysal/splot
 .. _fork: https://help.github.com/articles/fork-a-repo/
+
