### 2020-01-12 9847575
[DOC] add note about later Python version
