### 2020-01-14 80a052c
[DOC] add doc-links to 'Usage'
- of GSOC report
- of Scipy video
- of notebooks
