### 2020-01-14 58214d1
[BUG] format figure caption
- add figure caption with ** underneath iamges
