### 2020-01-17 c0ae8cd
[DOC] mention notebooks proj in README.md
and spell checking
