+++ README.md
@@ -13,15 +13,15 @@
 ## What is splot?
 
 `splot` connects spatial analysis done in [`PySAL`](https://github.com/pysal) to different popular visualization toolkits like [`matplotlib`](https://matplotlib.org).
-The `splot` package allows you to create both static plots ready for publication and interactive visualizations for quick iteration and spatial data exploration. The primary goal of `splot` is to enable you to visualize popular PySAL objects and gives you different views on your spatial analysis workflow.
+The `splot` package allows you to create both static plots ready for publication and interactive visualizations for quick iteration and spatial data exploration. The primary goal of `splot` is to enable you to visualize popular `PySAL` objects and gives you different views on your spatial analysis workflow.
 
-If you are new to splot and PySAL you will best get started with our [documentation](https://splot.readthedocs.io/en/latest/) and the short introduction [video](https://youtu.be/kriQOJMycIQ?t=2403) of the package at the Scipy 2018 conference!
+If you are new to `splot` and `PySAL` you will best get started with our [documentation](https://splot.readthedocs.io/en/latest/) and the short introduction [video](https://youtu.be/kriQOJMycIQ?t=2403) of the package at the Scipy 2018 conference!
 
 ## Installing splot
 
 ### Installing dependencies:
 
-splot is compatible with `Python` 3.6 and 3.7 and depends on `geopandas` 0.4.0 or later and `matplotlib` 2.2.2 or later.
+`splot` is compatible with `Python` 3.6 and 3.7 and depends on `geopandas` 0.4.0 or later and `matplotlib` 2.2.2 or later.
 
 splot also uses
 * `numpy`
@@ -29,10 +29,10 @@ splot also uses
 * `mapclassify`
 * `Ipywidgets`
 
-Depending on your spatial analysis workflow and the PySAL objects you would like to visualize, splot relies on:
+Depending on your spatial analysis workflow and the `PySAL` objects you would like to visualize, `splot` relies on:
 * PySAL 2.0
 
-or separate packages found in the PySAL stack:
+or separate packages found in the `PySAL` stack:
 * esda
 * libpysal
 * spreg
@@ -40,7 +40,7 @@ or separate packages found in the PySAL stack:
 
 ### Installing splot:
 
-There are two ways of accessing splot. First, splot is installed with the [PySAL 2.0](https://pysal.readthedocs.io/en/latest/installation.html) metapackage through:
+There are two ways of accessing `splot`. First, `splot` is installed with the [PySAL 2.0](https://pysal.readthedocs.io/en/latest/installation.html) metapackage through:
 
     $ pip install -U pysal
     
@@ -49,12 +49,12 @@ There are two ways of accessing splot. First, splot is installed with the [PySAL
     $ conda install -c conda-forge pysal
 
 
-Second, splot can be installed as a separate package. If you are using Anaconda, install splot via the conda utility:
+Second, `splot` can be installed as a separate package. If you are using Anaconda, install `splot` via the `conda` utility:
 
     $ conda install -c conda-forge splot
 
 
-Otherwise you can install splot from PyPI with pip:
+Otherwise you can install `splot` from `PyPI` with pip:
 
     $ pip install splot
 
@@ -63,13 +63,11 @@ Otherwise you can install splot from PyPI with pip:
 
 Usage examples for different spatial statistical workflows are provided as [notebooks](https://github.com/pysal/splot/tree/master/notebooks):
 * [for creating value-by-alpha maps](https://github.com/pysal/splot/blob/master/notebooks/mapping_vba.ipynb)
[…]
