### 2020-01-18 2090a4a
[REL] version 1.1.2 in __init__.py
