+++ splot/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.1.1"
+__version__ = "1.1.2"
 # __version__ has to be defined in the first line
 
 
