### 2020-01-18 pysal/splot#92 @slumnitz
solved in #97
