### 2020-01-18 pysal/splot#90 @slumnitz
solved in #97
