### 2020-01-18 pysal/splot#88 @slumnitz
addressed in #96
