### 2020-01-21 #1882 @slumnitz
@martinfleis @ResidentMario and @leouieda thank you for your review and comments. The new splot release includes solutions to the major documentation issues outlined. Currently as you noticed, splot tutorials require previous knowledge in spatial statistical analysis and are not aimed to teach statistical concepts.

Do you think a new tutorial like `a beginners guide to spatial statistical analysis in splot` would be suitable to fill this gap, or would you recommend referencing great blogposts, statistical literature is enough?
