### 2020-01-22 #1882 @martinfleis
@slumnitz the figure captions formatting is still strange. JOSS has its own style for captions, see the [last published paper](https://joss.theoj.org/papers/10.21105/joss.01800). Your figures seems to be encapsulated in links, which are not working anyway, causing captions not to render.
