### 2020-01-22 #1882 @martinfleis
@slumnitz to answer your question, I think good links are enough.
