### 2020-01-30 9f27bbf
Remove the links around figures in the JOSS paper
Include the links in the figure caption instead to avoid issues when 
compiling the paper PDF.
