### 2020-01-30 #1882 @leouieda
@slumnitz thank you keeping us in the loop. As @martinfleis said, the figures weren't rendering properly. I submitted a PR fixing this.

Also, we now have a compilation service that you can use to preview the paper (even on branches before merging PRs): https://whedon.theoj.org

@ResidentMario I see you still have the "Installation" review item unchecked. Could you please see if this has been addressed in the new version of splot?

@martinfleis you have the "Quality of writing" item unchecked. Has this been addressed in the updated paper?
