### 2020-01-30 #1882 @martinfleis
@leouieda I kept it unchecked due to figures formatting issue. Your PR resolves it, but note that you made a PR to your forked repo not the original repo.

I'll take it as resolved and finish my review.
