### 2020-01-31 #1882 @ResidentMario
I had the installation mark unchecked due to splot#88, but that's now been fixed so I went ahead and checked the box.

[The new docs](http://pysal.org/notebooks/viz/splot/intro.html) look great!

That's everything from me.
