### 2020-02-05 #1882 @leouieda
> Your PR resolves it, but note that you made a PR to your forked repo not the original repo.

Well, that was embarrassing :) Resolved now with pysal/splot#99

@martinfleis and @ResidentMario thank you for your reviews!

@slumnitz I'll go over the submission one more time to make any editorial suggestions that might be needed. After that, I'm happy to move forward with acceptance of your submission to JOSS :tada:
