### 2020-03-04 #1882 @ljwolf
The rerender of the paper after mering pysal/splot#99 is [here](https://res.cloudinary.com/hju22ue2k/image/upload/v1583329488/puu3xi4ibwtllubo20fg.pdf) and should conform to the requests, so let me or @slumnitz know if anything else is required.
