### 2020-03-17 #1882 @arfon
@ljwolf - I'm not sure of @leouieda's availability to handle the final steps here so...

At this point could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
