### 2020-03-17 #1882 @slumnitz
@ljwolf thanks for taking this on. I finally have a stable wifi connection again here in Italy. I will take this on tonight. Would you be able to review as soon as I made the PR's?
