### 2020-03-18 437941b
[REL] v1.1.3 in 'setup.py' and 'init.py'
