### 2020-03-18 6a67480
declaring package name and version variables
