### 2020-03-23 #1882 @slumnitz
@arfon thank you for helping. Splot has a new release v.1.1.3 on Zenodo

DOI: 10.5281/zenodo.3724199

I am not sure how I can update this thread. Maybe you could help? Thank you.
