+++ paper/paper.bib
@@ -174,6 +174,6 @@
   year = {2010},
   volume = 47,
   issue = 2,
-  pages = 130--140,
+  pages = {130--140},
   doi = {10.1179/000870409X124887534553372}
   }
