### 2020-03-23 #1882 @arfon
@slumnitz - this PR should fix the BibTeX issue here: https://github.com/pysal/splot/pull/104
