### 2020-03-23 #1882 @arfon
@slumnitz - could you take a look at the [DOI summary above](https://github.com/openjournals/joss-reviews/issues/1882#issuecomment-602641085) to check to see if the DOIs under `MISSING` are correct recommendations and also check the `INVALID` DOI?
