### 2020-03-23 61da587
[JOSS] doi reference correction
