+++ paper/paper.bib
@@ -9,7 +9,8 @@
    publisher = {{Riverhead Books}},
    author = {Johnson, Steven},
    month = oct,
-   year = {2007}
+   year = {2007},
+   doi = {10.1080/01944360802146329}
  }
 
  @incollection{arribas-bel2017looking,
@@ -70,7 +71,8 @@
    author = {Anselin, L. and Griffith, Daniel A},
    year = {1988},
    keywords = {Spatial dependence},
-   pages = {11--34}
+   pages = {11--34},
+   doi = {10.1111/j.1435-5597.1988.tb01155.x}
  }
 
  @book{anselin2014modern,
@@ -175,5 +177,5 @@
   volume = 47,
   issue = 2,
   pages = {130--140},
-  doi = {10.1179/000870409X124887534553372}
+  doi = {10.1179/000870409X12488753453372}
   }
