### 2020-03-23 #1882 @slumnitz
@arfon all missing doi's are added and the incorrect one was fixed.
